implementation of metrics for pointclouds and general
