+++ geomfum/_registry.py
@@ -352,8 +352,21 @@ class PointCloudPlotterRegistry(Registry):
     MAP = {}
 
 
-class HeatDistanceMetricRegistry(Registry):
+class _MeshHeatDistanceMetricRegistry(Registry):
+    has_internal = False
     MAP = {}
+
+
+class _PointSetHeatDistanceMetricRegistry(Registry):
+    has_internal = False
+    MAP = {}
+
+
+class HeatDistanceMetricRegistry(NestedRegistry):
+    Registries = {
+        True: _MeshHeatDistanceMetricRegistry,
+        False: _PointSetHeatDistanceMetricRegistry,
+    }
 
 
 def _create_register_funcs(module):
