+++ geomfum/metric/__init__.py
@@ -1,0 +1,1 @@
+from ._base import Metric, FinitePointSetMetric, HeatDistanceMetric
