+++ geomfum/metric/point_cloud.py.py
@@ -1,0 +1,1 @@
+"""Module containing metrics to calculate distances on a PointCloud."""
