+++ geomfum/wrap/__init__.py
@@ -10,10 +10,10 @@ from geomfum._registry import (
     register_landmark_wave_kernel_signature,
     register_laplacian_finder,
     register_mesh_plotter,
+    register_point_cloud_plotter,
     register_poisson_sampler,
     register_sinkhorn_neighbor_finder,
     register_wave_kernel_signature,
-    register_point_cloud_plotter,
 )
 from geomfum._utils import has_package
 
@@ -113,5 +113,13 @@ register_sinkhorn_neighbor_finder(
 )
 
 register_heat_distance_metric(
-    "pp3d", "Pp3dHeatDistanceMetric", requires="potpourri3d", as_default=True
+    True, "pp3d", "Pp3dMeshHeatDistanceMetric", requires="potpourri3d", as_default=True
+)
+
+register_heat_distance_metric(
+    False,
+    "pp3d",
+    "Pp3dPointSetHeatDistanceMetric",
+    requires="potpourri3d",
+    as_default=True,
 )
